// templates/web-app-react/src/api/generated/types.gen.ts
export type Item = {
    id: string;
    name: string;
    description: string | null;
    quantity: number;
    created_at: string;
    updated_at: string;
};